% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{earthquakes}
\alias{earthquakes}
\title{earthquakes}
\format{
An object of class \code{data.frame} with 49 rows and 4 columns.
}
\usage{
earthquakes
}
\description{
earthquakes
}
\keyword{datasets}
